* `NIntegrate` and auxiliary functions are moved from `mathics.core.atomic.numbers` to `mathics.core.numbers.calculus` * the non-member ``apply_N`` was moved to the new module ``mathics.core.evaluators``. The logic of ``mathics.builtin.atomic.numeric.N.n_apply_with_prec``   was also moved to that module. * most of flake8 errors on the changed modules was fixed.
