+++ mathics/builtin/compilation.py
@@ -10,7 +10,7 @@ When LLVM and Python libraries are available, compilation produces LLVM code.
 import ctypes
 
 
-from mathics.builtin.atomic.numbers import apply_N
+from mathics.core.evaluators import apply_N
 from mathics.builtin.base import Builtin
 from mathics.builtin.box.compilation import CompiledCodeBox
 from mathics.core.evaluation import Evaluation
