+++ mathics/builtin/drawing/graphics3d.py
@@ -6,7 +6,7 @@ Functions for working with 3D graphics.
 """
 
 
-from mathics.builtin.atomic.numbers import apply_N
+from mathics.core.evaluators import apply_N
 from mathics.builtin.base import Builtin
 from mathics.builtin.colors.color_directives import RGBColor
 from mathics.builtin.graphics import (
