+++ mathics/builtin/graphics.py
@@ -9,7 +9,7 @@ Drawing Graphics
 from math import sqrt
 
 
-from mathics.builtin.atomic.numbers import apply_N
+from mathics.core.evaluators import apply_N
 
 from mathics.builtin.base import (
     Builtin,
