+++ mathics/builtin/lists.py
@@ -25,7 +25,7 @@ from mathics.algorithm.clusters import (
     optimize,
 )
 
-from mathics.builtin.atomic.numbers import apply_N
+from mathics.core.evaluators import apply_N
 
 from mathics.builtin.base import (
     Builtin,
