+++ mathics/core/systemsymbols.py
@@ -26,6 +26,7 @@ SymbolLess = Symbol("Less")
 SymbolMachinePrecision = Symbol("MachinePrecision")
 SymbolMakeBoxes = Symbol("MakeBoxes")
 SymbolMinus = Symbol("Minus")
+SymbolN = Symbol("N")
 SymbolNumberQ = Symbol("NumberQ")
 SymbolNumericQ = Symbol("NumericQ")
 SymbolPattern = Symbol("Pattern")
